Fold deploy package into sdk Rename several files to their class name Rename TestCoreDeploy to TestCoreDeployer Add return types to more methods
