--- typescript/deploy/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@abacus-network/deploy",
-  "description": "Deploy tools for the Abacus network",
-  "version": "0.2.3",
-  "dependencies": {
-    "@abacus-network/app": "^0.2.3",
-    "@abacus-network/core": "^0.2.3",
-    "@abacus-network/sdk": "^0.2.3",
-    "@abacus-network/utils": "^0.2.3",
-    "@types/debug": "^4.1.7",
-    "cross-fetch": "^3.1.5",
-    "debug": "^4.3.4"
-  },
-  "devDependencies": {
-    "@types/node": "^16.9.1",
-    "ethers": "^5.6.8",
-    "prettier": "^2.4.1",
-    "ts-node": "^10.8.0",
-    "typescript": "^4.7.2"
-  },
-  "homepage": "https://www.useabacus.network",
-  "keywords": [
-    "Abacus",
-    "Typescript",
-    "Deployment"
-  ],
-  "license": "Apache-2.0",
-  "main": "dist/src/index.js",
-  "prepublish": "yarn build",
-  "repository": "https://github.com/abacus-network/abacus-monorepo",
-  "scripts": {
-    "build": "tsc",
-    "check": "tsc --noEmit",
-    "prettier": "prettier --write ./src"
-  },
-  "types": "dist/src/index.d.ts",
-  "files": [
-    "/dist"
-  ]
-}
