--- typescript/deploy/src/index.ts
@@ -1,21 +1,0 @@
-export { AbacusAppChecker, Ownable } from './check';
-export { CheckerViolation, EnvironmentConfig } from './config';
-export {
-  AbacusCoreDeployer,
-  CoreConfig,
-  ValidatorManagerConfig,
-} from './core/deploy';
-export { AbacusCoreChecker } from './core/check';
-export { AbacusDeployer } from './deploy';
-export { UpgradeBeaconViolation } from './proxy';
-export {
-  AbacusRouterChecker,
-  AbacusRouterDeployer,
-  RouterConfig,
-} from './router';
-export * as utils from './utils';
-export {
-  ContractVerifier,
-  getConstructorArguments,
-  VerificationInput,
-} from './verify';
