--- typescript/deploy/src/router/index.ts
@@ -1,3 +1,0 @@
-export { AbacusRouterChecker } from './check';
-export { AbacusRouterDeployer } from './deploy';
-export { RouterConfig } from './types';
