--- typescript/deploy/src/verify/index.ts
@@ -1,3 +1,0 @@
-export { ContractVerifier } from './ContractVerifier';
-export { ContractVerificationInput, VerificationInput } from './types';
-export { getConstructorArguments, getContractVerificationInput } from './utils';
