--- typescript/deploy/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "compilerOptions": {
-    "outDir": "./dist/",
-    "rootDir": "./"
-  },
-  "exclude": ["./node_modules/", "./dist/", "./tmp.ts"],
-  "extends": "../../tsconfig.json",
-  "include": [
-    "./*.ts",
-    "./config/**/*.ts",
-    "./scripts/**/*.ts",
-    "./src/**/*.ts",
-    "./test/**/*.ts"
-  ]
-}
