+++ typescript/sdk/package.json
@@ -3,10 +3,12 @@
   "description": "The official SDK for the Abacus Network",
   "version": "0.2.3",
   "dependencies": {
+    "@abacus-network/app": "^0.2.3",
     "@abacus-network/celo-ethers-provider": "^0.1.0",
     "@abacus-network/core": "^0.2.3",
     "@abacus-network/utils": "^0.2.3",
     "@types/debug": "^4.1.7",
+    "cross-fetch": "^3.1.5",
     "debug": "^4.3.4",
     "ethers": "^5.6.8"
   },
