--- typescript/sdk/src/core/index.ts
@@ -1,21 +1,0 @@
-export { AbacusCore, CoreContractsMap } from './app';
-export {
-  CoreContracts,
-  coreFactories,
-  InboxContracts,
-  OutboxContracts,
-} from './contracts';
-export { environments as coreEnvironments } from '../consts/environments';
-export {
-  AbacusLifecyleEvent,
-  AnnotatedDispatch,
-  AnnotatedLifecycleEvent,
-} from './events';
-export {
-  AbacusMessage,
-  AbacusStatus,
-  MessageStatus,
-  resolveDomain,
-  resolveId,
-  resolveNetworks,
-} from './message';
