+++ typescript/sdk/src/gas/calculator.ts
@@ -3,12 +3,12 @@ import { BigNumber, FixedNumber, ethers } from 'ethers';
 import { utils } from '@abacus-network/utils';
 
 import { chainMetadata } from '../consts/chainMetadata';
-import { AbacusCore } from '../core';
-import { MultiProvider } from '../provider';
+import { AbacusCore } from '../core/AbacusCore';
+import { MultiProvider } from '../providers/MultiProvider';
 import { ChainName, Remotes } from '../types';
+import { convertDecimalValue, mulBigAndFixed } from '../utils/number';
 
 import { DefaultTokenPriceGetter, TokenPriceGetter } from './token-prices';
-import { convertDecimalValue, mulBigAndFixed } from './utils';
 
 /**
  * A note on arithmetic:
