--- typescript/sdk/src/gas/index.ts
@@ -1,2 +1,0 @@
-export { InterchainGasCalculator } from './calculator';
-export { DefaultTokenPriceGetter, TokenPriceGetter } from './token-prices';
