+++ typescript/sdk/src/hardhat/TestCoreApp.ts
@@ -4,12 +4,12 @@ import { TestInbox, TestOutbox } from '@abacus-network/core';
 import { types, utils } from '@abacus-network/utils';
 
 import { chainMetadata } from '../consts/chainMetadata';
+import { AbacusCore } from '../core/AbacusCore';
 import {
-  AbacusCore,
   CoreContracts,
   InboxContracts,
   OutboxContracts,
-} from '../core';
+} from '../core/contracts';
 import { DomainIdToChainName } from '../domains';
 import { ProxiedContract } from '../proxy';
 import { ChainMap, ChainName, Remotes, TestChainNames } from '../types';
