--- typescript/sdk/src/hardhat/index.ts
@@ -1,2 +1,0 @@
-export * from './TestCoreApp';
-export * from './TestCoreDeploy';
