+++ typescript/sdk/src/providers/ChainConnection.ts
@@ -1,15 +1,6 @@
 import { ethers } from 'ethers';
 
-import { ChainMap, ChainName } from './types';
-import { MultiGeneric, objMap } from './utils';
-
-export interface IChainConnection {
-  provider: ethers.providers.Provider;
-  signer?: ethers.Signer;
-  overrides?: ethers.Overrides;
-  confirmations?: number;
-  blockExplorerUrl?: string;
-}
+import { IChainConnection } from '../types';
 
 export class ChainConnection {
   provider: ethers.providers.Provider;
@@ -40,24 +31,4 @@ export class ChainConnection {
       this.blockExplorerUrl
     }/address/${await this.signer!.getAddress()}`;
   }
-}
-
-export class MultiProvider<
-  Chain extends ChainName = ChainName,
-> extends MultiGeneric<Chain, ChainConnection> {
-  constructor(chainConnectionConfigs: ChainMap<Chain, IChainConnection>) {
-    super(
-      objMap(
-        chainConnectionConfigs,
-        (_, connection) => new ChainConnection(connection),
-      ),
-    );
-  }
-  getChainConnection(chain: Chain): ChainMap<Chain, ChainConnection>[Chain] {
-    return this.get(chain);
-  }
-  // This doesn't work on hardhat providers so we skip for now
-  // ready() {
-  //   return Promise.all(this.values().map((dc) => dc.provider!.ready));
-  // }
 }
