+++ typescript/sdk/src/utils/index.ts
@@ -1,7 +1,8 @@
 export * from './chains';
 export * from './ids';
 export * from './MultiGeneric';
+export * from './number';
 export * from './objects';
-export * from './RetryProvider';
+export * from './sets';
 export * from './signer';
 export * from './time';
